backend/requirements.txt
# Core requirements
gradio>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
requests>=2.31.0
euriai>=0.1.0
bcrypt>=4.1.2
PyJWT>=2.8.0

# PDF processing and text extraction
PyPDF2>=3.0.0

# Vector database and embeddings
faiss-cpu>=1.7.4

# Async support
aiohttp>=3.8.0

# Progress bars and utilities
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
